+++ lib/click_house/response/factory.rb
@@ -11,7 +11,10 @@ module ClickHouse
         return body if !body.is_a?(Hash) || !(body.key?('meta') && body.key?('data'))
 
         ResultSet.new(
-          meta: body.fetch('meta'), data: body.fetch('data'), totals: body['totals'], statistics: body['statistics']
+          meta: body.fetch('meta'),
+          data: body.fetch('data'),
+          totals: body['totals'],
+          statistics: body['statistics']
         )
       end
     end